keep visualzation name for enrichment
